+++ examples/eventually.dart
@@ -1,0 +1,16 @@
+import 'package:dartz/dartz.dart';
+import 'package:use_in_case/use_in_case.dart';
+
+final class Printer implements Interactor {
+  @override
+  Future<void> execute(Unit input) async {
+    print('Goodbye!');
+  }
+}
+
+void main() async {
+  final interactor = Printer();
+  await interactor
+      .eventually(() => print('Done.'))
+      .run(unit); // Hello!, Goodbye!
+}