+++ lib/src/after.dart
@@ -1,14 +1,15 @@
+import 'dart:async';
+
 import 'package:use_in_case/src/interactor.dart';
 
 extension After<Input, Output> on ParameterizedResultInteractor<Input, Output> {
   ParameterizedResultInteractor<Input, Output> after(
-    Future<void> Function(Output) callback,
+    FutureOr<void> Function(Output) callback,
   ) {
-    return InlinedParameterizedResultInteractor((input) {
-      return execute(input).then((output) async {
-        await callback(output);
-        return output;
-      });
+    return InlinedParameterizedResultInteractor((input) async {
+      final output = await execute(input);
+      await callback(output);
+      return output;
     });
   }
 }