+++ lib/src/eventually.dart
@@ -1,12 +1,18 @@
+import 'dart:async';
+
 import 'package:use_in_case/use_in_case.dart';
 
 extension Eventually<Input, Output>
     on ParameterizedResultInteractor<Input, Output> {
   ParameterizedResultInteractor<Input, Output> eventually(
-    Future<void> Function() callback,
+    FutureOr<void> Function() callback,
   ) {
-    return InlinedParameterizedResultInteractor((input) {
-      return execute(input).whenComplete(callback);
+    return InlinedParameterizedResultInteractor((input) async {
+      try {
+        return await execute(input);
+      } finally {
+        await callback();
+      }
     });
   }
 }