+++ lib/src/interactor.dart
@@ -1,7 +1,9 @@
+import 'dart:async';
+
 import 'package:dartz/dartz.dart';
 
 abstract interface class ParameterizedResultInteractor<Input, Output> {
-  Future<Output> execute(Input input);
+  FutureOr<Output> execute(Input input);
 }
 
 typedef ParameterizedInteractor<Input>
@@ -11,10 +13,10 @@ typedef Interactor = ParameterizedResultInteractor<Unit, void>;
 
 final class InlinedParameterizedResultInteractor<Input, Output>
     implements ParameterizedResultInteractor<Input, Output> {
-  final Future<Output> Function(Input input) _execute;
+  final FutureOr<Output> Function(Input input) _execute;
 
   const InlinedParameterizedResultInteractor(this._execute);
 
   @override
-  Future<Output> execute(Input input) => _execute(input);
+  FutureOr<Output> execute(Input input) => _execute(input);
 }