+++ lib/src/progress.dart
@@ -1,13 +1,15 @@
+import 'dart:async';
+
 import 'package:dartz/dartz.dart';
 import 'package:use_in_case/src/interactor.dart';
 
 abstract base class ParameterizedResultProgressInteractor<Input, Output,
     Progress> {
-  late final Future<void> Function(Progress progress) callback;
-  Future<Output> execute(Input input);
+  late final FutureOr<void> Function(Progress progress) callback;
+  FutureOr<Output> execute(Input input);
 
-  Future<void> emitProgress(Progress progress) {
-    return callback.call(progress);
+  FutureOr<void> emitProgress(Progress progress) {
+    return callback(progress);
   }
 }
 
@@ -21,9 +23,10 @@ typedef ProgressInteractor<Progress>
 extension ReceiveProgressExtension<Input, Output, Progress>
     on ParameterizedResultProgressInteractor<Input, Output, Progress> {
   ParameterizedResultInteractor<Input, Output> receiveProgress(
-    Future<void> Function(Progress) callback,
+    FutureOr<void> Function(Progress) callback,
   ) {
     this.callback = callback;
+
     return InlinedParameterizedResultInteractor<Input, Output>((input) {
       return execute(input);
     });